Staging

+++ src/Utils/Team/Roles/User.ts
@@ -58,6 +58,14 @@ async function removeUser({
     const roleRepository = getRepository(UserRoles);
     const role = await getUserRole({ user_id, team_id });
 
+    if (role.role.toLowerCase() === 'manager') {
+        throw new AppError({
+            message: 'You cannot remove a manager from team',
+            internalErrorCode: 41,
+            statusCode: 401,
+        });
+    }
+
     await roleRepository.remove(role);
 
     const cache = new Cache();
